assignment01/pyproject.toml
[project]
name = "assignment01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.8",
    "python-decouple>=3.8",
    "tavily-python>=0.7.11",
]
